+++ apps/klaviyo/src/app/api/webhooks/app-deleted/route.ts
@@ -1,14 +1,3 @@
-import { createAppDeletedHandler } from "@saleor/webhook-utils/app-deleted-handler";
-
-import { saleorApp } from "../../../../../saleor-app";
-import { createLogger } from "../../../../logger";
-
-const { handler, getWebhookManifest } = createAppDeletedHandler({
-  apl: saleorApp.apl,
-  logger: createLogger("APP_DELETED handler"),
-  webhookPath: "api/webhooks/app-deleted",
-});
-
-export { getWebhookManifest };
+import { handler } from "./webhook-definition";
 
 export const POST = handler;
