+++ apps/klaviyo/src/pages/api/manifest.ts
@@ -4,7 +4,7 @@ import { wrapWithLoggerContext } from "@saleor/apps-logger/node";
 import { withSpanAttributes } from "@saleor/apps-otel/src/with-span-attributes";
 
 import pkg from "../../../package.json";
-import { getWebhookManifest as getAppDeletedWebhookManifest } from "../../app/api/webhooks/app-deleted/route";
+import { getWebhookManifest as getAppDeletedWebhookManifest } from "../../app/api/webhooks/app-deleted/webhook-definition";
 import { env } from "../../env";
 import { loggerContext } from "../../logger-context";
 import { customerCreatedWebhook } from "./webhooks/customer-created";
